app/requirements.txt
gradio>=3.36.1
google-cloud-aiplatform==1.25.0
google-cloud-logging